Despagetthified code

+++ src/gui/ScreenProperties.java
@@ -1,6 +1,6 @@
 package gui;
 
 public class ScreenProperties {
-	public static double screenWidth = 1024,screenHeight = 768;
+	public static int screenWidth = 1024,screenHeight = 768;
 
 }
